pyproject.toml
[tool.poetry]
name = "aptos-blindness-detection"
version = "0.1.0"
description = "Kaggle APTOS 2019 Blindness Detection Solution"
authors = ["Gemini Agent <agent@example.com>"]
readme = "README.md"
packages = [{include = "src"}]

[tool.poetry.dependencies]
python = "^3.11"
tensorflow = [
    {version = "^2.16", markers = "sys_platform == 'darwin'"},
    {version = "^2.16", extras = ["and-cuda"], markers = "sys_platform == 'linux' or sys_platform == 'win32'"}
]
tensorflow-metal = {version = "^1.1", markers = "sys_platform == 'darwin'"}
keras = "^3.0"
pandas = "^2.2"
numpy = "^1.26"
opencv-python-headless = "^4.9"
albumentations = "^1.4"
hydra-core = "^1.3"
wandb = "^0.16"
optuna = "^3.5"
scikit-learn = "^1.4"
matplotlib = "^3.8"
seaborn = "^0.13"
pydot = "^2.0"

[tool.poetry.group.dev.dependencies]
ruff = "^0.3"
mypy = "^1.8"
pytest = "^8.0"
black = "^24.2"
isort = "^5.13"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 88
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "W", "B"]
ignore = []

[tool.mypy]
python_version = "3.11"
ignore_missing_imports = true
disallow_untyped_defs = true
check_untyped_defs = true
